Generate HTML using ascii doctor
Finding: Basically, Rest Doc asks to write up the whole API document anyways, and it's just a conduit of API document and unit tests. Not much time saving as you'd need to write everything. The learning curve is also steep and set up is cumbersome as you'd need to write up index.adoc.